requirements.txt
numpy
pandas
pathlib
opencv-python
tqdm
scikit-learn
ultralytics
PyYAML
